Add tests for GamesPlay commentController

+++ src/controllers/gamesPlay/commentController.js
@@ -4,6 +4,7 @@ import { authMiddleware } from "../../middlewares/authMiddleware.js";
 import { isAdmin } from "../../middlewares/isAdminMiddleware.js";
 
 import { commentDto } from "../../validators/gamesPlay/commentDto.js";
+import { mongooseIdDto } from "../../validators/mongooseIdDto.js";
 
 import { asyncErrorHandler } from "../../utils/errorUtils/asyncErrorHandler.js";
 import { CustomError } from "../../utils/errorUtils/customError.js";
@@ -16,6 +17,11 @@ export function commentController(commentService) {
         asyncErrorHandler(async (req, res) => {
             const gameId = req.params.gameId;
 
+            const { error: idError } = mongooseIdDto.validate({ id: gameId });
+            if (idError) {
+                throw new CustomError(idError.details[0].message, 400);
+            }
+
             const comments = await commentService.getAll(gameId);
 
             res.status(200).json(comments);
@@ -26,13 +32,18 @@ export function commentController(commentService) {
         "/",
         authMiddleware,
         asyncErrorHandler(async (req, res) => {
-            const { error } = commentDto.validate(req.body);
-            if (error) {
-                throw new CustomError(error.details[0].message, 400);
-            }
-
             const userId = req.user._id;
             const data = req.body;
+
+            const { error: idError } = mongooseIdDto.validate({ id: userId });
+            if (idError) {
+                throw new CustomError(idError.details[0].message, 400);
+            }
+
+            const { error: dataError } = commentDto.validate(data);
+            if (dataError) {
+                throw new CustomError(dataError.details[0].message, 400);
+            }
 
             const comment = await commentService.create(data, userId);
 
@@ -46,6 +57,13 @@ export function commentController(commentService) {
         isAdmin,
         asyncErrorHandler(async (req, res) => {
             const commentId = req.params.commentId;
+
+            const { error: idError } = mongooseIdDto.validate({
+                id: commentId,
+            });
+            if (idError) {
+                throw new CustomError(idError.details[0].message, 400);
+            }
 
             await commentService.remove(commentId);
